+++ tests/Support/lang/es/routes.php
@@ -3,5 +3,6 @@
 return [
     'example' => 'ejemplo',
     'products/{product}' => 'productos/{product}',
+    'posts/{post}' => 'articulos/{post}',
     'example.com' => 'es.example.com',
 ];